%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: AIC-methods.Rd 3122 2006-11-05 15:10:25Z leisch $
%
\name{AIC-methods}
\docType{methods}
\title{Methods for Function AIC}
\alias{AIC,flexmix-method}
\alias{AIC,stepFlexmix-method}
\description{Compute the Akaike Information Criterion.}
\section{Methods}{
  \describe{
    \item{object = flexmix:}{Compute the AIC of a \code{flexmix} object}
    \item{object = stepFlexmix:}{Compute the AIC of all
      models contained in the \code{stepFlexmix} object.}
  }
}
\keyword{methods}
