\name{candy}
\alias{candy}
\docType{data}
\title{Candy packs purchased}
\description{
  The data is from a new product and concept test where the number of
  individual packs of hard candy purchased within the past 7 days is recorded.
}
\usage{data("candy")}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{Packages}}{a numeric vector}
    \item{\code{Freq}}{a numeric vector}
  }
}
\source{
  D. Boehning, E. Dietz, P. Schlattmann (1998): Recent Developments in
  Computer-Assisted Analysis of Mixtures. Biometrics 54(2), 525--536.
}
\references{  
  J. Magidson, J. K. Vermunt (2004): Latent Class Models. In
  D. W. Kaplan (ed.), The Sage Handbook of Quantitative Methodology for
  the Social Sciences, 175--198. Thousand Oakes: Sage Publications.

  D. Boehning, E. Dietz, P. Schlattmann (1998): Recent Developments in
  Computer-Assisted Analysis of Mixtures. Biometrics 54(2), 525--536.

  W. R. Dillon, A. Kumar (1994): Latent structure and other mixture
  models in marketing: An integrative survey and overview. In
  R. P. Bagozzi (ed.), Advanced methods of marketing research,
  352--388. Cambridge, UK: Blackwell.
}

\keyword{datasets}
