\name{flexmix-internal}
\alias{FLXglmmodel-class}
\alias{FLXrefit-class}
\alias{FLXrefitglm-class}
\alias{summary.flexmix-class}
\title{Internal FlexMix functions}
\description{
  Internal flexmix functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
