\name{group}
\docType{methods}
\alias{group}
\alias{group-methods}
\alias{group,flexmix-method}
\alias{group,FLXM-method}
\alias{group,FLXMRglmfix-method}
\title{Extract Grouping Variable}
\description{
  Extract grouping variable for all observations.
}
\usage{
\S4method{group}{flexmix}(object)
\S4method{group}{FLXM}(object)
\S4method{group}{FLXMRglmfix}(object)
}
\arguments{
  \item{object}{an object of class \code{flexmix}.}
}
\keyword{methods}
\author{Bettina Gruen}
