%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: logLik-methods.Rd 1664 2005-06-13 06:11:03Z leisch $
%
\name{logLik-methods}
\docType{methods}
\title{Methods for Function logLik in Package `flexmix'}
\alias{logLik,flexmix-method}
\description{Evaluate the log-likelihood.  This function is defined as
  an S4 generic in the \code{stats4} package.}
\section{Methods}{
  \describe{
    \item{object = flexmix}{Evaluate the log-likelihood of an
      \code{flexmix} object}
  }
}
\keyword{methods}
