%
%  Copyright (C) 2004-2011 Friedrich Leisch and Bettina Gruen
%  $Id: AIC-methods.Rd 4666 2011-02-23 15:52:35Z gruen $
%
\name{AIC-methods}
\docType{methods}
\title{Methods for Function AIC}
\alias{AIC,flexmix-method}
\alias{AIC,stepFlexmix-method}
\description{Compute the Akaike Information Criterion.}
\section{Methods}{
  \describe{
    \item{object = flexmix:}{Compute the AIC of a \code{flexmix} object}
    \item{object = stepFlexmix:}{Compute the AIC of all
      models contained in the \code{stepFlexmix} object.}
  }
}
\keyword{methods}
