\name{AIC-methods}
\docType{methods}
\title{Methods for Function AIC}
\alias{AIC-methods}
\alias{AIC,ANY-method}
\alias{AIC,flexmix-method}
\description{Evaluate the Akaike Information Criterion.
  This function is defined as an old-style generic in the \code{base}
  package.  In this package is it converted to an S4 generic.}
\section{Methods}{
  \describe{
    \item{object = ANY}{Evaluate the AIC of a general object}
    \item{object = flexmix}{Evaluate the AIC of an
      \code{flexmix} object}
  }
}
\keyword{methods}
