%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: AIC-methods.Rd 1664 2005-06-13 06:11:03Z leisch $
%
\name{AIC-methods}
\docType{methods}
\title{Methods for Function AIC}
\alias{AIC,flexmix-method}
\description{Compute the Akaike Information Criterion.}
\section{Methods}{
  \describe{
    \item{object = flexmix:}{Compute the AIC of a \code{flexmix} object}
  }
}
\keyword{methods}
