%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: BIC-methods.Rd 1664 2005-06-13 06:11:03Z leisch $
%
\name{BIC-methods}
\docType{methods}
\title{Methods for Function BIC}
\alias{BIC,flexmix-method}
\description{Compute the Bayesian Information Criterion.}
\section{Methods}{
  \describe{
    \item{object = flexmix:}{Compute the BIC of a \code{flexmix} object}
  }
}
\keyword{methods}
