%
%  Copyright (C) 2004-2008 Friedrich Leisch and Bettina Gruen
%  $Id: BIC-methods.Rd 3912 2008-03-13 15:10:24Z gruen $
%
\name{BIC-methods}
\docType{methods}
\title{Methods for Function BIC}
\alias{BIC,flexmix-method}
\alias{BIC,stepFlexmix-method}
\description{Compute the Bayesian Information Criterion.}
\section{Methods}{
  \describe{
    \item{object = flexmix:}{Compute the BIC of a \code{flexmix} object}
    \item{object = stepFlexmix:}{Compute the BIC of all
      models contained in the \code{stepFlexmix} object.}
  }
}
\keyword{methods}
