%
%  Copyright (C) 2004-2015 Friedrich Leisch and Bettina Gruen
%  $Id: candy.Rd 5008 2015-01-13 20:30:25Z gruen $
%
\name{candy}
\alias{candy}
\docType{data}
\title{Candy Packs Purchased}
\description{
  The data is from a new product and concept test where the number of
  individual packs of hard candy purchased within the past 7 days is recorded.
}
\usage{data("candy")}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{Packages}}{a numeric vector}
    \item{\code{Freq}}{a numeric vector}
  }
}
\source{
  D. Boehning, E. Dietz and P. Schlattmann. Recent Developments in
  Computer-Assisted Analysis of Mixtures. Biometrics 54(2), 525--536,
  1998.
}
\references{  
  J. Magidson and J. K. Vermunt. Latent Class Models. In
  D. W. Kaplan (ed.), The Sage Handbook of Quantitative Methodology for
  the Social Sciences, 175--198, 2004. Thousand Oakes: Sage Publications.

  D. Boehning, E. Dietz and P. Schlattmann. Recent Developments in
  Computer-Assisted Analysis of Mixtures. \emph{Biometrics},
  \bold{54}(2), 525--536, 1998.

  W. R. Dillon and A. Kumar. Latent structure and other mixture
  models in marketing: An integrative survey and overview. In
  R. P. Bagozzi (ed.), Advanced methods of marketing research,
  352--388, 1994. Cambridge, UK: Blackwell.
}

\keyword{datasets}
