%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: flexmix-internal.Rd 1664 2005-06-13 06:11:03Z leisch $
%
\name{flexmix-internal}
\alias{FLXglmmodel-class}
\alias{FLXrefit-class}
\alias{FLXrefitglm-class}
\alias{summary.flexmix-class}
\title{Internal FlexMix functions}
\description{
  Internal flexmix functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
