%
%  Copyright (C) 2004-2008 Friedrich Leisch and Bettina Gruen
%  $Id: flexmix-internal.Rd 4493 2009-12-21 16:12:54Z gruen $
%
\name{flexmix-internal}
\alias{FLXgetModelmatrix}
\alias{FLXcheckComponent}
\alias{FLXcheckComponent,FLXM-method}
\alias{FLXcheckComponent,FLXMRfix-method}
\alias{FLXdeterminePostunscaled}
\alias{FLXdeterminePostunscaled,FLXM-method}
\alias{FLXdeterminePostunscaled,FLXMRfix-method}
\alias{FLXgetK}
\alias{FLXgetK,FLXM-method}
\alias{FLXgetK,FLXMRfix-method}
\alias{FLXgetModelmatrix}
\alias{FLXgetModelmatrix,FLXM-method}
\alias{FLXgetModelmatrix,FLXMRfix-method}
\alias{FLXgetModelmatrix,FLXMRrobglm-method}
\alias{FLXgetModelmatrix,FLXMRziglm-method}
\alias{FLXgetModelmatrix,FLXP-method}
\alias{FLXgetObs}
\alias{FLXgetObs,FLXM-method}
\alias{FLXgetObs,FLXMRfix-method}
\alias{FLXmstep}       
\alias{FLXmstep,FLXM-method}
\alias{FLXmstep,FLXMRfix-method}
\alias{FLXmstep,FLXMRrobglm-method}
\alias{FLXmstep,FLXMRziglm-method}
\alias{FLXremoveComponent}
\alias{FLXremoveComponent,FLXM-method}
\alias{FLXremoveComponent,FLXMRfix-method}
\alias{FLXremoveComponent,FLXMRrobglm-method}
\alias{FLXremoveComponent,FLXMRziglm-method}
\alias{FLXMRglm-class}
\alias{FLXR-class}
\alias{FLXRMRglm-class}
\alias{FLXRPmultinom-class}
\alias{summary.flexmix-class}
\alias{posterior,FLXM,listOrdata.frame-method}
\alias{FLXMRfix-class}
\alias{FLXMRglmfix-class}
\alias{FLXRMRglmfix-class}
\alias{predict,FLXMRglmfix-method}
\alias{fitted,FLXMRglmfix-method}
\alias{summary,FLXRMRglmfix-method}
\alias{listOrdata.frame-class}
\title{Internal FlexMix functions}
\description{
  Internal flexmix functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
