%
%  Copyright (C) 2004-2008 Friedrich Leisch and Bettina Gruen
%  $Id: flexmix-internal.Rd 4638 2010-11-22 14:01:52Z gruen $
%
\name{flexmix-internal}
\alias{FLXgetModelmatrix}
\alias{FLXcheckComponent}
\alias{FLXcheckComponent,FLXM-method}
\alias{FLXcheckComponent,FLXMRfix-method}
\alias{FLXdeterminePostunscaled}
\alias{FLXdeterminePostunscaled,FLXM-method}
\alias{FLXdeterminePostunscaled,FLXMRfix-method}
\alias{FLXgetK}
\alias{FLXgetK,FLXM-method}
\alias{FLXgetK,FLXMRfix-method}
\alias{FLXgetModelmatrix}
\alias{FLXgetModelmatrix,FLXM-method}
\alias{FLXgetModelmatrix,FLXMRfix-method}
\alias{FLXgetModelmatrix,FLXMRmgcv-method}
\alias{FLXgetModelmatrix,FLXMRrobglm-method}
\alias{FLXgetModelmatrix,FLXMRziglm-method}
\alias{FLXgetModelmatrix,FLXP-method}
\alias{FLXgetObs}
\alias{FLXgetObs,FLXM-method}
\alias{FLXgetObs,FLXMRfix-method}
\alias{FLXmstep}       
\alias{FLXmstep,FLXM-method}
\alias{FLXmstep,FLXMRfix-method}
\alias{FLXmstep,FLXMRmgcv-method}
\alias{FLXmstep,FLXMRrobglm-method}
\alias{FLXmstep,FLXMRziglm-method}
\alias{FLXremoveComponent}
\alias{FLXremoveComponent,FLXM-method}
\alias{FLXremoveComponent,FLXMRfix-method}
\alias{FLXremoveComponent,FLXMRrobglm-method}
\alias{FLXremoveComponent,FLXMRziglm-method}
\alias{FLXMRglm-class}
\alias{FLXR-class}
\alias{FLXRMRglm-class}
\alias{FLXRPmultinom-class}
\alias{summary.flexmix-class}
\alias{posterior,FLXM,listOrdata.frame-method}
\alias{FLXMRfix-class}
\alias{FLXMRglmfix-class}
\alias{FLXRMRglmfix-class}
\alias{predict,FLXMRglmfix-method}
\alias{fitted,FLXMRglmfix-method}
\alias{summary,FLXRMRglmfix-method}
\alias{listOrdata.frame-class}
\alias{refit_optim}
\alias{refit_optim,FLXM-method}
\alias{refit_optim,FLXMC-method}
\alias{refit_optim,FLXMRglm-method}
\alias{refit_optim,FLXMRziglm-method}
\alias{refit_optim,FLXP-method}
\alias{FLXgetDesign}
\alias{FLXgetDesign,FLXM-method}
\alias{FLXgetDesign,FLXMRglmfix-method}
\alias{FLXgetDesign,FLXMRziglm-method}
\alias{FLXgetParameters}
\alias{FLXgetParameters,FLXdist-method}
\alias{FLXgetParameters,FLXM-method}
\alias{FLXgetParameters,FLXMC-method}
\alias{FLXgetParameters,FLXMRglm-method}
\alias{FLXgetParameters,FLXP-method}
\alias{FLXgetParameters,FLXPmultinom-method}
\alias{FLXreplaceParameters}
\alias{FLXreplaceParameters,FLXdist-method}
\alias{FLXreplaceParameters,FLXM-method}
\alias{FLXreplaceParameters,FLXMC-method}
\alias{FLXreplaceParameters,FLXMRglm-method}
\alias{FLXreplaceParameters,FLXP-method}
\alias{FLXreplaceParameters,FLXPmultinom-method}
\alias{FLXgradlogLikfun}
\alias{FLXgradlogLikfun,flexmix-method}
\alias{FLXgradlogLikfun,FLXM-method}
\alias{FLXgradlogLikfun,FLXMRglm-method}
\alias{FLXgradlogLikfun,FLXP-method}
\alias{existGradient}
\alias{existGradient,FLXM-method}
\alias{existGradient,FLXMRglm-method}
\alias{existGradient,FLXMRglmfix-method}
\alias{existGradient,FLXP-method}

\title{Internal FlexMix functions}
\description{
  Internal flexmix functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
