%
%  Copyright (C) 2004-2005 Friedrich Leisch
%  $Id: flexmix-internal.Rd 3221 2007-01-22 14:27:40Z gruen $
%
\name{flexmix-internal}
\alias{FLXgetModelmatrix}
\alias{FLXcheckComponent}
\alias{FLXcheckComponent,FLXM-method}
\alias{FLXcheckComponent,FLXMRglmfix-method}
\alias{FLXdeterminePostunscaled}
\alias{FLXdeterminePostunscaled,FLXM-method}
\alias{FLXdeterminePostunscaled,FLXMRglmfix-method}
\alias{FLXgetK}
\alias{FLXgetK,FLXM-method}
\alias{FLXgetK,FLXMRglmfix-method}
\alias{FLXgetModelmatrix}
\alias{FLXgetModelmatrix,FLXM-method}
\alias{FLXgetModelmatrix,FLXMRglmfix-method}
\alias{FLXgetModelmatrix,FLXMRziglm-method}
\alias{FLXgetModelmatrix,FLXP-method}
\alias{FLXgetObs}
\alias{FLXgetObs,FLXM-method}
\alias{FLXgetObs,FLXMRglmfix-method}
\alias{FLXmstep}       
\alias{FLXmstep,FLXM-method}
\alias{FLXmstep,FLXMRglmfix-method}
\alias{FLXmstep,FLXMRziglm-method}
\alias{FLXremoveComponent}
\alias{FLXremoveComponent,FLXM-method}
\alias{FLXremoveComponent,FLXMRglmfix-method}
\alias{FLXremoveComponent,FLXMRziglm-method}
\alias{FLXMRglm-class}
\alias{FLXR-class}
\alias{FLXRMRglm-class}
\alias{FLXRPmultinom-class}
\alias{summary.flexmix-class}
\alias{posterior,FLXM,listOrdata.frame-method}
\alias{FLXMRglmfix-class}
\alias{FLXRMRglmfix-class}
\alias{predict,FLXMRglmfix-method}
\alias{fitted,FLXMRglmfix-method}
\alias{refit,FLXMRglmfix-method}
\alias{summary,FLXRMRglmfix-method}
\alias{listOrdata.frame-class}
\title{Internal FlexMix functions}
\description{
  Internal flexmix functions, methods and classes.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
