%
%  Copyright (C) 2004-2008 Friedrich Leisch and Bettina Gruen
%  $Id: logLik-methods.Rd 3912 2008-03-13 15:10:24Z gruen $
%
\name{logLik-methods}
\docType{methods}
\title{Methods for Function logLik in Package `flexmix'}
\alias{logLik,flexmix-method}
\description{Evaluate the log-likelihood.  This function is defined as
  an S4 generic in the \code{stats4} package.}
\section{Methods}{
  \describe{
    \item{object = flexmix}{Evaluate the log-likelihood of an
      \code{flexmix} object}
  }
}
\keyword{methods}
