\name{logLik-methods}
\docType{methods}
\title{Methods for Function logLik}
\alias{logLik,flexmix-method}
\description{Evaluate the log-likelihood.  This function is defined as
  an old-style generic in the \code{base} package.  In this package is
  it converted to an S4 generic.}
\section{Methods}{
  \describe{
    \item{object = ANY}{Evaluate the log-likelihood of a general object}
    \item{object = flexmix}{Evaluate the log-likelihood of an
      \code{flexmix} object}
  }
}
\keyword{methods}
