%
%  Copyright (C) 2004-2015 Friedrich Leisch and Bettina Gruen
%  $Id: logLik-methods.Rd 5008 2015-01-13 20:30:25Z gruen $
%
\name{logLik-methods}
\docType{methods}
\title{Methods for Function logLik in Package \pkg{flexmix}}
\alias{logLik,flexmix-method}
\alias{logLik,stepFlexmix-method}
\description{Evaluate the log-likelihood.  This function is defined as
  an S4 generic in the \code{stats4} package.}
\section{Methods}{
  \describe{
    \item{object = flexmix}{Evaluate the log-likelihood of an
      \code{flexmix} object}
  }
}
\keyword{methods}
